third_party/blink/web_tests/virtual/customizable-select-disabled/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-option-arrow-scroll.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] topofdocument: Arrow keys on options should not scroll the document.
  assert_equals: First option should be focused when opening the select. expected Element node <option>one</option> but got Element node <select id="topofdocument">\n  <option>one</option>\n  <opt...
[FAIL] bottomofdocument: Arrow keys on options should not scroll the document.
  assert_equals: First option should be focused when opening the select. expected Element node <option>one</option> but got Element node <select id="bottomofdocument">\n  <option>one</option>\n  <...
Harness: the test ran to completion.
